requirements.txt
flask
psycopg2-binary
python-dotenv
gunicorn
requests
beautifulsoup4